Add `ntops.torch.pooling._calculate_output_size`
